Add files via upload
Business Problem:
Spotify, is a digital music, podcast, and video service that gives access to millions of songs and other content creators all over the world. Spotify aims to provide more successful songs to its platform in order to draw in and keep more customers in light of the growing competition and demand. To determine whether a song will become a hit or not, users are looking to use the dataset and to utilize machine learning and deep learning models to access data resources and prioritise the efforts for prediction. We therefore set out to find which characteristics are the best indicators of a hit song and to concentrate on those characteristics while curating Spotify's collections list. We also set out to establish the threshold probability of a hit song to be used when choosing new songs for the platform.

About the data:
This dataset consists of features for tracks (1960's - 2010's) fetched using Spotify's Web API. The tracks are labeled '1' or '0' ('Hit' or 'Flop'). This dataset can be used to make a classification model that predicts whether a track would be a 'Hit' or not.

Note: 'Flop' here merely implies that it is not a song that probably could not be considered popular in the mainstream. It does not objectively considers a track inferior, bad or a failure if its labeled 'Flop'.

Data Understanding (features/labels):
track: The Name of the track.

artist: The Name of the Artist.

uri: The uniform resource identifier for the track.

danceability: Danceability describes how suitable a track is for dancing based on a combination of musical elements including tempo, rhythm stability, beat strength, and overall regularity. A value of 0.0 is least danceable and 1.0 is most danceable.

energy: Energy is a measure from 0.0 to 1.0 and represents a perceptual measure of intensity and activity. Typically, energetic tracks feel fast, loud, and noisy. For example, death metal has high energy, while a Bach prelude scores low on the scale. Perceptual features contributing to this attribute include dynamic range, perceived loudness, timbre, onset rate, and general entropy.

key: The estimated overall key of the track. Integers map to pitches using standard Pitch Class notation. E.g. 0 = C, 1 = C?/D?, 2 = D, and so on. If no key was detected, the value is -1.

loudness: The overall loudness of a track in decibels (dB). Loudness values are averaged across the entire track and are useful for comparing relative loudness of tracks. Loudness is the quality of a sound that is the primary psychological correlate of physical strength (amplitude). Values typical range between -60 and 0 db.

mode: Mode indicates the modality (major or minor) of a track, the type of scale from which its melodic content is derived. Major is represented by 1 and minor is 0.

speechiness: Speechiness detects the presence of spoken words in a track. The more exclusively speech-like the recording (e.g. talk show, audio book, poetry), the closer to 1.0 the attribute value. Values above 0.66 describe tracks that are probably made entirely of spoken words. Values between 0.33 and 0.66 describe tracks that may contain both music and speech, either in sections or layered, including such cases as rap music. Values below 0.33 most likely represent music and other non-speech-like tracks.

acousticness: A confidence measure from 0.0 to 1.0 of whether the track is acoustic. 1.0 represents high confidence the track is acoustic. The distribution of values for this feature look like this:

instrumentalness: Predicts whether a track contains no vocals. “Ooh” and “aah” sounds are treated as instrumental in this context. Rap or spoken word tracks are clearly “vocal”. The closer the instrumentalness value is to 1.0, the greater likelihood the track contains no vocal content. Values above 0.5 are intended to represent instrumental tracks, but confidence is higher as the value approaches 1.0. The distribution of values for this feature look like this:

liveness: Detects the presence of an audience in the recording. Higher liveness values represent an increased probability that the track was performed live. A value above 0.8 provides strong likelihood that the track is live.

valence: A measure from 0.0 to 1.0 describing the musical positiveness conveyed by a track. Tracks with high valence sound more positive (e.g. happy, cheerful, euphoric), while tracks with low valence sound more negative (e.g. sad, depressed, angry).

tempo: The overall estimated tempo of a track in beats per minute (BPM). In musical terminology, tempo is the speed or pace of a given piece and derives directly from the average beat duration.

duration_ms: The duration of the track in milliseconds.

time_signature: An estimated overall time signature of a track. The time signature (meter) is a notational convention to specify how many beats are in each bar (or measure).

chorus_hit: This the the author's best estimate of when the chorus would start for the track. Its the timestamp of the start of the third section of the track. This feature was extracted from the data received by the API call for Audio Analysis of that particular track.

sections: The number of sections the particular track has. This feature was extracted from the data received by the API call for Audio Analysis of that particular track.

target: The target variable for the track. It can be either '0' or '1'. '1' implies that this song has featured in the weekly list (Issued by Billboards) of Hot-100 tracks in that decade at least once and is therefore a 'hit'. '0' Implies that the track is a 'flop'.

Data Processing:
Dataset being thoroughly checked for null and duplicate counts and successfully addressed them.
Dataset has mixed types of features (numerical/catogorical/discrete).
Visualized the data for its patterns and skewness by implementing various exploratory data analysis methods.
Project aimed at building classification model with a binary outcome of '1' or '0' ('Hit' or 'Flop').
Data Modeling:
Several models were attempted in addressing this problem, increasing in complexity.
First, to get a sense for what a untrained model would perform like, we used a dummy classifier (baseline model) to make predictions according to the majority class of the training data.
Second, iterated through several Machine Learning models (Random Forest, Logistic Regression, XGBoost etc)
Third, implemented Deep Learning models (Convolutional Neural Network, Multilayer Perceptron - MLP, Dense Neural Network etc).
Tools/applications used for this were Google collab and Jupyter Notebook.
Evaluation:
Models performance was evaluated using accuracy and ROC-AUC scores.
Implemented feature importance and shaply to determine the importance of the data features.